+++ toonalyst/src/main/java/com/toonalyst/domain/webtoon/WebtoonDTO.java
@@ -27,7 +27,7 @@ import lombok.ToString;
 @ToString
 public class WebtoonDTO {
 	private String platForm;	//제공플랫폼
-	private int titleId;		//웹툰 고유 ID 숫자로 구성됨
+	private long titleId;		//웹툰 고유 ID 숫자로 구성됨
 	private String titleName;	//웹툰 제목
 	private String writer;		//작가
 	private float rating;		//웹툰 제공 플랫폼 별점
